feat(backend/avaliacao): implementa endpoint de escrever avaliação

+++ backend/DTOs/Produto/AvaliacaoDTO.cs
@@ -5,7 +5,7 @@ namespace backend.DTOs.Produto;
 public class AvaliacaoDTO
 {
     public int Id { get; set; }
-    public int Nota { get; set; }
+    public float Nota { get; set; }
     public string Comentario { get; set; } = string.Empty;
 
     public ProdutoDTO? Produto { get; set; }
